requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.0.post1 